requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
python-jose[cryptography]
passlib[bcrypt]
pydantic
httpx
authlib
